Move endpoint to _synapse
