--- synapse/rest/identity/__init__.py
@@ -1,28 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2019 The Matrix.org Foundation C.I.C.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from synapse.http.server import JsonResource
-from synapse.rest.identity.threepid_submit_token_servlet import (
-    ThreepidSubmitTokenServlet,
-)
-
-
-class IdentityRestResource(JsonResource):
-    """The REST resource which gets mounted at /_matrix/identity/api/v1"""
-
-    def __init__(self, hs):
-        JsonResource.__init__(self, hs, canonical_json=False)
-
-        ThreepidSubmitTokenServlet(hs).register(self)
